+++ pycosio/__init__.py
@@ -1,17 +1,17 @@
 # coding=utf-8
 """Python Cloud Object Storage I/O"""
 
-__version__ = '1.0.0a1'
+__version__ = '1.0.0'
 
 # Adds names to public interface
 # Shadowing "open" built-in name is done to provides "pycosio.open" function
 from pycosio._core.std_functions import (
-    cos_open as open, copy, getmtime, getsize, isfile, listdir, relpath)
+    cos_open as open, copy, getmtime, getsize, isfile, relpath)
 from pycosio._core.storage_manager import register
 
 __all__ = [
     # Standard functions
-    'open', 'copy', 'getmtime', 'getsize', 'listdir', 'relpath', 'isfile',
+    'open', 'copy', 'getmtime', 'getsize', 'relpath', 'isfile',
 
     # Utilities
     'register']