+++ tests/test_http.py
@@ -6,7 +6,8 @@ from wsgiref.handlers import format_date_time
 
 import pytest
 
-from tests.utilities import parse_range, check_head_methods, check_raw_read_methods
+from tests.utilities import (
+    parse_range, check_head_methods, check_raw_read_methods)
 
 
 def test_handle_http_errors():
@@ -59,6 +60,7 @@ def test_http_raw_io():
 
     class Response:
         """Fake response"""
+
         def __init__(self):
             self.status_code = 200
 
@@ -70,6 +72,7 @@ def test_http_raw_io():
 
     class Session:
         """Fake Session"""
+
         def __init__(self, *_, **__):
             """Do nothing"""